+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 
 ### Changed
 
+- (API) `GameRules` is changed to an interface.
 - Breaking block related warnings are now moved to debug channel.
 - Improved code readability for I18n module.
 - NBT library is now updated to 3.0.10.